refactor: adapting the territories, blocks, and streets pages to accept the new way of accessing by ID in the URLs and updating custom hooks to push to the correct URL
