+++ src/common/block/type.ts
@@ -1,25 +1,24 @@
-
 export type IUseBlock = {
-   block: IBlock
-   actions: IActions
-}
+  block: IBlock;
+  actions: IActions;
+};
 
 export type IActions = {
-   goToStreet: (streetId: number) => Promise<void>
-}
+  goToStreet: (streetId: number) => void;
+};
 
 export type IBlock = {
-   territoryId: number
-   territoryName: string
-   blockId: number
-   blockName: string
-   addresses: IAddress[]
-}
+  territoryId: number;
+  territoryName: string;
+  blockId: number;
+  blockName: string;
+  addresses: IAddress[];
+};
 
 export type IAddress = {
-   id: number
-   name: string
-   houses: House[]
-}
+  id: number;
+  name: string;
+  houses: House[];
+};
 
-export type House = string
+export type House = string;
