+++ src/common/territory/useTerritory.ts
@@ -20,9 +20,7 @@ export const useTerritory = (territoryId: number, initialState?: ITerritory) => 
   const [isLoading, setIsLoading] = useState<Mode>('loading');
 
   const getTerritories = useCallback(async (id: number): Promise<void> => {
-    console.log({ id })
     if (!id) {
-      setIsLoading('not-found');
       return;
     }
     const { status, data } = await TerritoryGateway.in().getById(id);
